utilizando user por empresa

+++ Backend/src/modules/user/services/DeleteUser.service.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from 'tsyringe';
+import AppError from '../../../shared/errors/AppError';
 import IUserRepository from '../interfaces/IUserRepository';
 
 
@@ -9,7 +10,12 @@ export class DeleteUserService {
         private repository: IUserRepository
     ) { }
 
-    public async execute(id: number): Promise<void> {
+    public async execute(id: number, idCompany:number): Promise<void> {
+        const user = await this.repository.findByID(id,idCompany)
+
+        if(!user)
+            throw new AppError('User not found')
+
         await this.repository.delete(id);
     }
